requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
django-htmx==1.17.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
